Table: Reorganize shared code to make it possible for LogsTable to omit some options (#117201)
